Cleaned up Filter, added desktop styling
