+++ packages/esm-patient-test-results-app/src/filter/filter-set.scss
@@ -2,38 +2,21 @@
 @import "~carbon-components/src/globals/scss/vars";
 @import "~carbon-components/src/globals/scss/mixins";
 
-.floatingRightButton {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  z-index: 9;
-}
-.expandButton {
-  background-color: white;
-  border: 1px solid black;
-}
-
-.floatingRightButton > .expandButton > svg {
-  margin: 0;
-}
-
-
-.filterContainer {
+// background of filter, and spacing between containers
+.nestedAccordion {
   background-color: $openmrs-background-grey;
   margin: $spacing-02 0;
+  @media (min-width: $breakpoint-desktop-min) {
+    background-color: $ui-background;
+  }
 }
 
-.filterNode {
-  padding: 1rem;
-}
-
-.filterItem {
-  padding: 0.5rem 1rem 0.5rem 4rem;
-}
-
+// our special accordion rules
 .nestedAccordion > :global(.bx--accordion--start) > :global(.bx--accordion__item--active) {
   border-left: 0.3rem solid var(--brand-01);
-  margin: 1.5rem 0;
+  @media (max-width: $breakpoint-tablet-max) {
+    margin: 1.5rem 0;
+  }
 }
 
 .nestedAccordion :global {
@@ -54,10 +37,22 @@
 
   .bx--accordion__title {
     margin: 0 0 0 0.8rem;
+    @media (min-width: $breakpoint-desktop-min) {
+      margin: 0 0 0 0.5rem;
+    }
   }
 
   .bx--accordion__arrow {
     margin: 0.4rem 0 0 1rem;
+    @media (min-width: $breakpoint-desktop-min) {
+      margin: 0.4rem 0 0 0.6rem;
+    }
+  }
+
+  @media (min-width: $breakpoint-desktop-min) {
+    .bx--accordion__item > button {
+      padding: 0;
+    }
   }
 
   // Chevron transformations
@@ -76,3 +71,9 @@
   }
 }
 
+.filterItem {
+  padding: 0.5rem 1rem 0.5rem 4rem;
+  @media (min-width: $breakpoint-desktop-min) {
+    padding: 0.2rem 0 0.2rem 4rem;
+  }
+}
